Move out flow extension methods to FlowExtensions.kt
